refactor: rename EQ DSP module to biquad EQ
