+++ src/lib/dsp/biquad-eq-worklet.ts
@@ -1,4 +1,4 @@
-import { type EqParameters, BiquadEq } from "./eq.ts";
+import { type EqParameters, BiquadEq } from "./biquad-eq.ts";
 
 const PROCESSOR_NAME = "biquad-eq";
