+++ src/lib/recorder/audio-buffer-playback.ts
@@ -1,5 +1,5 @@
 import { BiquadEqNode } from "../dsp/biquad-eq-node.ts";
-import type { EqParameters } from "../dsp/eq.ts";
+import type { EqParameters } from "../dsp/biquad-eq.ts";
 import { createPitchShifterNode } from "../dsp/pitch-shifter-node.ts";
 import type {
   AudioContextTransport,
